--- features/weather/src/main/kotlin/com/neillon/weather/data/WeatherService.kt
@@ -1,9 +1,0 @@
-package com.neillon.weather.data
-
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface WeatherService {
-    @GET("current.json")
-    suspend fun search(@Query("q") query: String): ApiResponse
-}
