+++ features/weather/src/main/kotlin/com/neillon/weather/data/service/WeatherApiError.kt
@@ -1,13 +1,13 @@
-package com.neillon.weather.data
+package com.neillon.weather.data.service
 
 import com.google.gson.annotations.SerializedName
 
 // 400 - BAD request errors
-data class ApiError(
-    @SerializedName("error") val error: ApiErrorData
+data class WeatherApiError(
+    @SerializedName("error") val error: WeatherApiErrorData
 )
 
-data class ApiErrorData(
+data class WeatherApiErrorData(
     @SerializedName("code") val code: Int,
     @SerializedName("message") val message: String
 )
