+++ features/weather/src/main/kotlin/com/neillon/weather/data/service/WeatherService.kt
@@ -1,0 +1,9 @@
+package com.neillon.weather.data.service
+
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface WeatherService {
+    @GET("current.json")
+    suspend fun search(@Query("q") query: String): WeatherApiResponse
+}
