+++ features/weather/src/main/kotlin/com/neillon/weather/domain/WeatherRepository.kt
@@ -1,5 +1,5 @@
 package com.neillon.weather.domain
 
 interface WeatherRepository {
-    suspend fun search(query: String): Result<Weather>
+    suspend fun search(query: String): Weather
 }
